feat: Add support for .tfvars file (#11)
Included in the changes:

The ability to merge simple key value files, like .tfvars which is a small subset of HCL
DRYs up how the merging and formatting works
Slight change in formatting. Where 2 new lines would get added at EOF, this would be a single new line at EOF. However, in the 1 case where there was 1 new line, this removes it. This is preferable IMO and we should keep it as such until the formatting (of a single new line) is consistent across all iterations.

BREAKING CHANGE: Although not breaking any contract/interface, I've marked this change as breaking change since the formatting of the return values has changed.